+++ tests/test006_bf_vdif_utils.py
@@ -1,11 +1,11 @@
 from mwax_mover.mwax_bf_vdif_utils import stitch_vdif_files_and_write_hdr
+from tests_common import setup_test_directories
 import pytest
 import os
 
-output_dir = "tests/data/test006"
-
 
 def test_stitch_zero_files():
+    output_dir = setup_test_directories("test006")
 
     filenames = []
     metafits_filename = ""
@@ -15,6 +15,7 @@ def test_stitch_zero_files():
 
 
 def test_stitch_one_file():
+    output_dir = setup_test_directories("test006")
 
     filenames = [
         "tests/data/1454343736/1454343736_1454343736_ch109_beam00.vdif",
@@ -30,14 +31,15 @@ def test_stitch_one_file():
         metafits_filename, filenames, output_dir
     )
 
-    assert output_vdif_filename == f"{output_dir}/1454343736_ch109_beam00.vdif"
-    assert output_hdr_filename == f"{output_dir}/1454343736_ch109_beam00.hdr"
+    assert output_vdif_filename == os.path.join(output_dir, "1454343736_ch109_beam00.vdif")
+    assert output_hdr_filename == os.path.join(output_dir, "1454343736_ch109_beam00.hdr")
 
     assert os.path.exists(output_vdif_filename)
     assert os.path.exists(output_hdr_filename)
 
 
 def test_stitch_many_files2():
+    output_dir = setup_test_directories("test006")
 
     filenames = [
         "tests/data/1454343736/1454343736_1454343736_ch109_beam01.vdif",
@@ -54,8 +56,8 @@ def test_stitch_many_files2():
         metafits_filename, filenames, output_dir
     )
 
-    assert output_vdif_filename == f"{output_dir}/{obs_id}_ch109_beam01.vdif"
-    assert output_hdr_filename == f"{output_dir}/{obs_id}_ch109_beam01.hdr"
+    assert output_vdif_filename == os.path.join(output_dir, f"{obs_id}_ch109_beam01.vdif")
+    assert output_hdr_filename == os.path.join(output_dir, f"{obs_id}_ch109_beam01.hdr")
 
     assert os.path.exists(output_vdif_filename)
     assert os.path.exists(output_hdr_filename)