+++ tests/test007_bf_filterbank_utils.py
@@ -1,3 +1,4 @@
+from tests_common import setup_test_directories
 from mwax_mover.mwax_bf_filterbank_utils import (
     stitch_filterbank_files,
     get_stitched_filename,
@@ -5,8 +6,6 @@ from mwax_mover.mwax_bf_filterbank_utils import (
 )
 import pytest
 import os
-
-output_dir = "tests/data/test007"
 
 
 def test_get_stitched_filename():
@@ -21,6 +20,7 @@ def test_get_stitched_filename():
 
 
 def test_stitch_zero_files():
+    output_dir = setup_test_directories("test007")
 
     filenames = []
 
@@ -29,6 +29,7 @@ def test_stitch_zero_files():
 
 
 def test_stitch_one_file():
+    output_dir = setup_test_directories("test007")
 
     filenames = [
         "tests/data/1451758560/1451758560_1451758560_ch109_beam00.fil",
@@ -36,12 +37,13 @@ def test_stitch_one_file():
 
     output_filterbank_filename = stitch_filterbank_files(filenames, output_dir)
 
-    assert output_filterbank_filename == "tests/data/test007/1451758560_ch109_beam00.fil"
+    assert output_filterbank_filename == os.path.join(output_dir, "1451758560_ch109_beam00.fil")
 
     assert os.path.exists(output_filterbank_filename)
 
 
 def test_stitch_many_files():
+    output_dir = setup_test_directories("test007")
 
     filenames = [
         "tests/data/1451758560/1451758560_1451758560_ch109_beam00.fil",
@@ -56,7 +58,7 @@ def test_stitch_many_files():
 
     output_filterbank_filename = stitch_filterbank_files(filenames, output_dir)
 
-    assert output_filterbank_filename == "tests/data/test007/1451758560_ch109_beam00.fil"
+    assert output_filterbank_filename == os.path.join(output_dir, "1451758560_ch109_beam00.fil")
     assert os.path.exists(output_filterbank_filename)
 
     # Get new data len- check it is the same as 1+2